Refaktorer BrevmalService og legg til filtrering av brevbakerbrev
