+++ gradle.properties
@@ -11,7 +11,7 @@ jupiterVersion=5.10.3
 hamkrestVersion=1.8.0.1
 logstashVersion=7.4
 micrometerVersion=1.12.5
-apiModelVersion=94
+apiModelVersion=95
 jacksonJsr310Version=2.17.0
 mockkVersion=1.13.10
 exposedVersion=0.55.0
