+++ pensjon-brevbaker/src/main/kotlin/no/nav/pensjon/brev/maler/redigerbar/BekreftelsePaaFlyktningstatus.kt
@@ -1,9 +1,9 @@
 package no.nav.pensjon.brev.maler.redigerbar
 
+import no.nav.pensjon.brev.api.model.Sakstype
 import no.nav.pensjon.brev.api.model.TemplateDescription
 import no.nav.pensjon.brev.api.model.maler.Brevkode
 import no.nav.pensjon.brev.api.model.maler.EmptyRedigerbarBrevdata
-import no.nav.pensjon.brev.maler.fraser.common.Felles
 import no.nav.pensjon.brev.maler.fraser.common.Felles.fulltNavn
 import no.nav.pensjon.brev.template.Language.Bokmal
 import no.nav.pensjon.brev.template.RedigerbarTemplate
@@ -20,13 +20,14 @@ import no.nav.pensjon.brevbaker.api.model.FoedselsnummerSelectors.value
 import no.nav.pensjon.brevbaker.api.model.LetterMetadata
 import no.nav.pensjon.brevbaker.api.model.NAVEnhetSelectors.navn
 
-
 @TemplateModelHelpers
 object BekreftelsePaaFlyktningstatus : RedigerbarTemplate<EmptyRedigerbarBrevdata> {
 
+    // PE_IY_03_049
     override val kode = Brevkode.Redigerbar.PE_BEKREFTELSE_PAA_FLYKTNINGSTATUS
-    override val kategori: TemplateDescription.Brevkategori =
-        TemplateDescription.Brevkategori.INNHENTE_OPPLYSNINGER
+    override val kategori = TemplateDescription.Brevkategori.INNHENTE_OPPLYSNINGER
+    override val brevkontekst = TemplateDescription.Brevkontekst.SAK
+    override val sakstyper = Sakstype.all
 
 
     override val template = createTemplate(
