+++ pensjon-brevbaker/src/main/kotlin/no/nav/pensjon/brev/maler/redigerbar/InformasjonOmSaksbehandlingstid.kt
@@ -1,5 +1,6 @@
 package no.nav.pensjon.brev.maler.redigerbar
 
+import no.nav.pensjon.brev.api.model.Sakstype
 import no.nav.pensjon.brev.api.model.TemplateDescription
 import no.nav.pensjon.brev.api.model.maler.Brevkode
 import no.nav.pensjon.brev.api.model.maler.redigerbar.InformasjonOmSaksbehandlingstidDto
@@ -25,8 +26,12 @@ import no.nav.pensjon.brevbaker.api.model.LetterMetadata
 
 @TemplateModelHelpers
 object InformasjonOmSaksbehandlingstid : RedigerbarTemplate<InformasjonOmSaksbehandlingstidDto> {
+
+    // AP_INFO_STID_MAN (MF 000130)
     override val kode = Brevkode.Redigerbar.INFORMASJON_OM_SAKSBEHANDLINGSTID
     override val kategori = TemplateDescription.Brevkategori.INFORMASJONSBREV
+    override val brevkontekst = TemplateDescription.Brevkontekst.SAK
+    override val sakstyper = Sakstype.all
 
     override val template = createTemplate(
         name = kode.name,
