+++ pensjon-brevbaker/src/main/kotlin/no/nav/pensjon/brev/maler/redigerbar/VarselOmMuligAvslag.kt
@@ -1,5 +1,6 @@
 package no.nav.pensjon.brev.maler.redigerbar
 
+import no.nav.pensjon.brev.api.model.Sakstype
 import no.nav.pensjon.brev.api.model.TemplateDescription
 import no.nav.pensjon.brev.api.model.maler.Brevkode
 import no.nav.pensjon.brev.api.model.maler.EmptyRedigerbarBrevdata
@@ -14,9 +15,11 @@ import no.nav.pensjon.brevbaker.api.model.LetterMetadata
 @TemplateModelHelpers
 object VarselOmMuligAvslag : RedigerbarTemplate<EmptyRedigerbarBrevdata> {
 
-
+    // PE_IY_03_051
     override val kode = Brevkode.Redigerbar.PE_VARSEL_OM_MULIG_AVSLAG
     override val kategori: TemplateDescription.Brevkategori = TemplateDescription.Brevkategori.VARSEL
+    override val brevkontekst = TemplateDescription.Brevkontekst.SAK
+    override val sakstyper = Sakstype.all
 
     override val template = createTemplate(
         name = kode.name,
