+++ skribenten-backend/src/main/kotlin/no/nav/pensjon/brev/skribenten/model/Brevmal.kt
@@ -1,0 +1,337 @@
+package no.nav.pensjon.brev.skribenten.model
+
+import no.nav.pensjon.brev.api.model.Sakstype
+import no.nav.pensjon.brev.api.model.Sakstype.*
+import no.nav.pensjon.brev.api.model.TemplateDescription
+import no.nav.pensjon.brev.api.model.TemplateDescription.Brevkategori
+import no.nav.pensjon.brev.skribenten.services.BrevdataDto
+import no.nav.pensjon.brev.skribenten.services.SpraakKode
+import no.nav.pensjon.brevbaker.api.model.LanguageCode
+
+enum class BrevSystem { EXSTREAM, DOKSYS, BREVBAKER }
+
+interface LetterMetadata {
+    val brevkode: String
+    val brevkontekst: TemplateDescription.Brevkontekst?
+    val isForSakskontekst: Boolean get() = brevkontekst in setOf(TemplateDescription.Brevkontekst.SAK, TemplateDescription.Brevkontekst.ALLE)
+    val isForVedtakskontekst: Boolean get() = brevkontekst in setOf(TemplateDescription.Brevkontekst.VEDTAK, TemplateDescription.Brevkontekst.ALLE)
+    val isRedigerbart: Boolean
+
+    fun isForSakstype(sakstype: Sakstype): Boolean
+    fun isRelevantRegelverk(sakstype: Sakstype, forGammeltRegelverk: Boolean): Boolean
+    fun toApi(): Api.Brevmal
+
+    /**
+     * Brevmetadata om brevmaler fra pensjon-brevmetadata (Exstream/Doksys)
+     */
+    data class Legacy(val data: BrevdataDto, private val hasSakstype: Sakstype) : LetterMetadata {
+        override val brevkode: String get() = data.brevkodeIBrevsystem
+        override val brevkontekst: TemplateDescription.Brevkontekst?
+            get() = when(data.brevkontekst) {
+                BrevdataDto.BrevkontekstCode.ALLTID -> TemplateDescription.Brevkontekst.ALLE
+                BrevdataDto.BrevkontekstCode.SAK -> TemplateDescription.Brevkontekst.SAK
+                BrevdataDto.BrevkontekstCode.VEDTAK -> TemplateDescription.Brevkontekst.VEDTAK
+                null -> null
+            }
+        override val isRedigerbart: Boolean get() = data.redigerbart
+        override fun isForSakstype(sakstype: Sakstype) = sakstype == hasSakstype
+
+        override fun isRelevantRegelverk(sakstype: Sakstype, forGammeltRegelverk: Boolean): Boolean =
+            if (data.brevregeltype != null) {
+                when (sakstype) {
+                    ALDER -> if (forGammeltRegelverk) {
+                        data.brevregeltype.gjelderGammeltRegelverk()
+                    } else {
+                        data.brevregeltype.gjelderNyttRegelverk()
+                    }
+
+                    UFOREP -> data.brevregeltype.gjelderGammeltRegelverk()
+                    BARNEP, AFP, AFP_PRIVAT, FAM_PL, GAM_YRK, GENRL, GJENLEV, GRBL, KRIGSP, OMSORG -> true
+                }
+            } else {
+                true
+            }
+
+        override fun toApi(): Api.Brevmal = with(data) {
+            Api.Brevmal(
+                name = dekode,
+                id = brevkodeIBrevsystem,
+                spraak = sprak ?: emptyList(),
+                brevsystem = when (brevsystem) {
+                    BrevdataDto.BrevSystem.DOKSYS -> BrevSystem.DOKSYS
+                    BrevdataDto.BrevSystem.GAMMEL -> BrevSystem.EXSTREAM
+                },
+                brevkategori = kategoriOverrides[brevkodeIBrevsystem]?.toKategoriTekst() ?: this.brevkategori?.toKategoriTekst(),
+                dokumentkategoriCode = this.dokumentkategori,
+                redigerbart = redigerbart,
+                redigerbarBrevtittel = isRedigerbarBrevtittel(),
+            )
+        }
+
+        private fun BrevdataDto.BrevkategoriCode.toKategoriTekst() =
+            when (this) {
+                BrevdataDto.BrevkategoriCode.BREV_MED_SKJEMA -> "Brev med skjema"
+                BrevdataDto.BrevkategoriCode.INFORMASJON -> "Informasjon"
+                BrevdataDto.BrevkategoriCode.INNHENTE_OPPL -> "Innhente opplysninger"
+                BrevdataDto.BrevkategoriCode.NOTAT -> "Notat"
+                BrevdataDto.BrevkategoriCode.OVRIG -> "Øvrig"
+                BrevdataDto.BrevkategoriCode.VARSEL -> "Varsel"
+                BrevdataDto.BrevkategoriCode.VEDTAK -> "Vedtak"
+            }
+    }
+
+    /**
+     * Brevmetadata om brevmaler fra Brevbaker
+     */
+    data class Brevbaker(val data: TemplateDescription.Redigerbar) : LetterMetadata {
+        override fun toApi(): Api.Brevmal = with(data) {
+            Api.Brevmal(
+                name = metadata.displayTitle,
+                id = name,
+                brevsystem = BrevSystem.BREVBAKER,
+                spraak = this.languages.map { it.toSpraakKode() },
+                brevkategori = kategori.toKategoriTekst(),
+                dokumentkategoriCode = metadata.brevtype.toDokumentkategoriCode(),
+                redigerbart = true,
+                redigerbarBrevtittel = false,
+            )
+        }
+
+        override val brevkontekst: TemplateDescription.Brevkontekst get() = data.brevkontekst
+        override val isRedigerbart: Boolean = true
+        override val brevkode: String get() = data.name
+        override fun isForSakstype(sakstype: Sakstype) = sakstype in data.sakstyper
+        override fun isRelevantRegelverk(sakstype: Sakstype, forGammeltRegelverk: Boolean) = true
+
+        private fun LanguageCode.toSpraakKode(): SpraakKode =
+            when (this) {
+                LanguageCode.BOKMAL -> SpraakKode.NB
+                LanguageCode.NYNORSK -> SpraakKode.NN
+                LanguageCode.ENGLISH -> SpraakKode.EN
+            }
+
+        private fun no.nav.pensjon.brevbaker.api.model.LetterMetadata.Brevtype.toDokumentkategoriCode(): BrevdataDto.DokumentkategoriCode =
+            when (this) {
+                no.nav.pensjon.brevbaker.api.model.LetterMetadata.Brevtype.VEDTAKSBREV -> BrevdataDto.DokumentkategoriCode.VB
+                no.nav.pensjon.brevbaker.api.model.LetterMetadata.Brevtype.INFORMASJONSBREV -> BrevdataDto.DokumentkategoriCode.IB
+            }
+    }
+}
+
+private fun Brevkategori.toKategoriTekst() =
+    when (this) {
+        Brevkategori.ETTEROPPGJOER -> "Etteroppgjør"
+        Brevkategori.FOERSTEGANGSBEHANDLING -> "Førstegangsbehandling"
+        Brevkategori.VEDTAK_ENDRING_OG_REVURDERING -> "Vedtak - endring og revurdering"
+        Brevkategori.VEDTAK_FLYTTE_MELLOM_LAND -> "Vedtak - flytte mellom land"
+        Brevkategori.SLUTTBEHANDLING -> "Sluttbehandling"
+        Brevkategori.INFORMASJONSBREV -> "Informasjonsbrev"
+        Brevkategori.VARSEL -> "Varsel"
+        Brevkategori.VEDTAK_EKSPORT -> "Vedtak - eksport"
+        Brevkategori.OMSORGSOPPTJENING -> "Omsorgsopptjening"
+        Brevkategori.UFOEREPENSJON -> "Uførepensjon"
+        Brevkategori.INNHENTE_OPPLYSNINGER -> "Innhente opplysninger"
+        Brevkategori.LEVEATTEST -> "Leveattest"
+        Brevkategori.FEILUTBETALING -> "Feilutbetaling"
+        Brevkategori.KLAGE_OG_ANKE -> "Klage og anke"
+        Brevkategori.POSTERINGSGRUNNLAG -> "Posteringsgrunnlag"
+        Brevkategori.FRITEKSTBREV -> "Fritekstbrev"
+    }
+
+private val kategoriOverrides: Map<String, Brevkategori> = mapOf(
+    Brevkategori.ETTEROPPGJOER to listOf(
+        "PE_AF_03_101",
+        "PE_AF_04_100",
+        "PE_AF_04_101",
+        "PE_AF_04_102",
+        "PE_AF_04_103",
+        "PE_AF_04_104",
+        "PE_AF_04_105",
+        "PE_AF_04_106",
+        "PE_AF_04_108",
+        "PE_AF_04_109",
+        "PE_UT_04_400",
+        "PE_UT_04_401",
+        "PE_UT_04_402"
+    ),
+    Brevkategori.FOERSTEGANGSBEHANDLING to listOf(
+        "AP_AVSL_TIDLUTTAK",
+        "AP_AVSL_UTTAK",
+        "AP_INNV_AVT_MAN",
+        "AP_INNV_MAN",
+        "BP_AVSL_MAN",
+        "PE_AF_04_001",
+        "PE_AF_04_010",
+        "PE_AF_04_111",
+        "PE_AF_04_112",
+        "PE_AP_04_001",
+        "PE_AP_04_010",
+        "PE_AP_04_202",
+        "PE_AP_04_203",
+        "PE_AP_04_211",
+        "PE_AP_04_212",
+        "PE_AP_04_901",
+        "PE_AP_04_902",
+        "PE_AP_04_910",
+        "PE_AP_04_911",
+        "PE_AP_04_913",
+        "PE_AP_04_914",
+        "PE_BP_04_001",
+        "PE_BP_04_002",
+        "PE_BP_04_010",
+        "PE_BP_04_011",
+        "PE_GP_04_001",
+        "PE_GP_04_010",
+        "PE_GP_04_023",
+        "PE_GP_04_024",
+        "PE_GP_04_026",
+        "PE_GP_04_031",
+        "PE_GP_04_033",
+        "PE_UT_04_101",
+        "PE_UT_04_104",
+        "PE_UT_04_117",
+        "PE_UT_04_118"
+    ),
+    Brevkategori.VEDTAK_ENDRING_OG_REVURDERING to listOf(
+        "AP_AVSL_ENDR",
+        "AP_AVSL_FT_MAN",
+        "AP_AVSL_GJRETT_MAN",
+        "AP_ENDR_EPS_MAN",
+        "AP_ENDR_FT_MAN",
+        "AP_ENDR_GJRETT_MAN",
+        "AP_ENDR_GRAD_MAN",
+        "AP_ENDR_INST_MAN",
+        "AP_ENDR_OPPTJ_MAN",
+        "AP_ENDR_STANS_MAN",
+        "AP_OPPH_FT_MAN",
+        "BP_OPPH_MAN",
+        "PE_AF_04_020",
+        "PE_AF_04_114",
+        "PE_AP_04_020",
+        "PE_AP_04_214",
+        "PE_AP_04_215",
+        "PE_AP_04_216",
+        "PE_AP_04_220",
+        "PE_AP_04_920",
+        "PE_BP_04_020",
+        "PE_BP_04_021",
+        "PE_BP_04_031",
+        "PE_FT_01_002",
+        "PE_FT_01_003",
+        "PE_FT_01_006",
+        "PE_FT_01_007",
+        "PE_GP_04_020",
+        "PE_GP_04_028",
+        "PE_GP_04_029",
+        "PE_GP_04_030",
+        "PE_UT_04_100",
+        "PE_UT_04_102",
+        "PE_UT_04_109",
+        "PE_UT_04_114",
+        "PE_UT_06_300",
+        "PE_UT_07_100"
+    ),
+    Brevkategori.VEDTAK_FLYTTE_MELLOM_LAND to listOf(
+        "AP_ENDR_FLYTT_MAN",
+        "AP_STANS_FLYTT_MAN",
+        "PE_AP_04_223",
+        "PE_AP_04_224",
+        "PE_AP_04_225",
+        "PE_IY_04_125",
+        "PE_IY_04_126",
+        "PE_IY_04_127"
+    ),
+    Brevkategori.SLUTTBEHANDLING to listOf(
+        "INFO_P1",
+        "PE_AP_04_903",
+        "PE_AP_04_904",
+        "PE_AP_04_912",
+        "PE_GP_04_025",
+        "PE_GP_04_032",
+        "PE_UT_04_106",
+        "PE_UT_04_107"
+    ),
+    Brevkategori.INFORMASJONSBREV to listOf(
+        "AP_INFO_STID_MAN",
+        "DOD_INFO_RETT_MAN",
+        "PE_AP_04_922",
+        "PE_BP_01_001",
+        "PE_BP_01_002",
+        "PE_IY_03_163",
+        "PE_IY_03_164",
+        "PE_IY_03_169",
+        "PE_IY_03_170",
+        "PE_IY_03_171",
+        "PE_IY_03_172",
+        "PE_IY_03_173",
+        "PE_IY_03_174",
+        "PE_IY_03_175",
+        "PE_IY_03_180",
+        "PE_IY_05_041",
+        "PE_UP_07_105",
+        "PE_UT_04_001",
+        "PE_UT_04_004"
+    ),
+    Brevkategori.VARSEL to listOf("VARSEL_REVURD", "PE_IY_03_051", "PE_IY_03_179"),
+    Brevkategori.VEDTAK_EKSPORT to listOf("PE_GP_04_022", "PE_UT_04_103", "PE_UT_04_115"),
+    Brevkategori.OMSORGSOPPTJENING to listOf("OMSORG_EGEN_MAN", "PE_IY_04_010"),
+    Brevkategori.UFOEREPENSJON to listOf(
+        "PE_UP_04_001",
+        "PE_UP_04_020",
+        "PE_UP_04_010",
+        "PE_UP_04_021",
+        "PE_UP_04_022",
+        "PE_UP_04_023",
+        "PE_UP_04_024",
+        "PE_UP_04_025",
+        "PE_UP_04_026",
+        "PE_UP_04_027",
+        "PE_UP_04_028",
+        "PE_UP_04_029",
+        "PE_UP_04_030",
+        "PE_UP_07_010",
+        "PE_UT_04_300"
+    ),
+    Brevkategori.INNHENTE_OPPLYSNINGER to listOf(
+        "PE_UP_07_100",
+        "PE_UT_04_003",
+        "HENT_INFO_MAN",
+        "PE_UT_04_002",
+        "PE_IY_03_167",
+        "PE_IY_03_047",
+        "PE_IY_03_048",
+        "PE_IY_03_168",
+        "PE_IY_03_178",
+        "PE_IY_05_006",
+        "PE_IY_05_007",
+        "PE_IY_05_008",
+        "PE_IY_03_049",
+        "PE_GP_01_010"
+    ),
+    Brevkategori.LEVEATTEST to listOf("PE_IY_03_176", "PE_IY_03_177", "PE_IY_05_411", "PE_IY_05_510", "PE_IY_05_410", "PE_IY_05_511"),
+    Brevkategori.FEILUTBETALING to listOf(
+        "VARSEL_TILBAKEBET",
+        "VEDTAK_TILBAKEKREV",
+        "VEDTAK_TILBAKEKREV_MIDL",
+        "PE_IY_04_060",
+        "PE_IY_04_061",
+        "PE_IY_05_027"
+    ),
+    Brevkategori.KLAGE_OG_ANKE to listOf(
+        "PE_IY_03_151",
+        "PE_IY_03_152",
+        "PE_IY_03_158",
+        "PE_IY_03_159",
+        "PE_IY_03_160",
+        "PE_IY_04_050",
+        "PE_IY_04_051",
+        "PE_IY_03_150",
+        "PE_IY_03_153",
+        "PE_IY_03_154",
+        "PE_IY_03_157",
+        "PE_IY_03_161",
+        "PE_IY_03_162"
+    ),
+    Brevkategori.POSTERINGSGRUNNLAG to listOf("PE_OK_06_100", "PE_OK_06_101", "PE_OK_06_102"),
+    Brevkategori.FRITEKSTBREV to listOf("PE_IY_03_156", "PE_IY_05_300"),
+).flatMap { kategori -> kategori.value.map { it to kategori.key } }.toMap()
